Tricky situaton:
import_fixture calls :environment so that solrizer will have all of your models defined.  In the case of hydra-head, this sets Rails.root to the test host's root.  As a result, after importing one fixture, it looks in the wrong place for the subsequent fixture files.

Solution: remove the line in uva.rake that tries to use Rails.root to set the fixture path.

... also fixed a bug in the puts statement in import_fixture