Moving model and job behavior to separate gem
The underlying models and many of the jobs of sufia are a useful
foundation for other applications. This is most evident when the routes
and controller "workflow" needs to vary by a substantial amount.

The sufia-models gem is presently packaged inside of the sufia
repository, but is something that would change at a later date. This is
done as a middle-step for full extraction.

These changes do not represent the complete separation of models/jobs
from the sufia gem, but should be considered a functional first step.
The underlying sufia tests continue to pass, as do Scholarsphere,
the Curate gem (https://github.com/ndlib/curate), and CurateND.