Moving resque rake task to sufia-models
This morning we spent several hours tracking down an issue related to
our resque-pool. As it turns out, resque-pool has a default task for
'resque:pool:setup'; However, without the task defined at the project
level, the rails environment is not loaded. This in turn means that
the config/initializers/redis_config.rb does not get loaded. Without
the config loaded, resque-pool assumes redis is running on the,
localhost which was not our case.