Add proxy deposit, sticky proxies, and transfers of ownership.
These three intertwined features allow a user to transfer ownership of files to other users, to deposit on behalf of other users, and to authorize other users to deposit on their behalf.  They were ported from ScholarSphere, which explains why 1) some specs use the older "should"-style rspec tests, and 2) the feature specs follow some slightly different conventions.  We should ultimately align ScholarSphere and Sufia in this area, but in the interest of incorporating these features sooner rather than later, I suggest that we do this in a separate branch.

Note that some of the changes in this commit were not necessary for proxies and transfers to work properly but rather to get the test suite to run and to pass -- I was not able to discern what in the codebase changed that necessitated these changes (primarily to the feature specs and the LocalAuthority model spec, but also some others).