Move the Collection model into the Sufia namespace
Many engines may define a Collection model, so we don't want to pollute
the namespace.  For backwards compatibility there is a Collection model
generated into the application's directory that extends the
Sufia::Collection.  This ensures that we don't need to update the
hasModel assertions in the Fedora repository.

As this is a breaking change (if the Collection model is not generated
into the application), it's recommended that this PR target the next
major revision of Sufia (5.0)