Change the way we log depositor change events
We had a ticket in ScholarSphere to make this change many months ago but it was never prioritized. Now that proxies and transfers have moved into Sufia, I thought it might be a good time to put this into place. In short, we want depositor change events to be logged publicly for the proxy depositor (who initiated the change) but logged to the dashboard of the depositor as a sort of notification, and we log the event only to the followers of the depositor, who now 'owns' the deposit, rather than the proxy depositor's followers