Noids should not be second-class citizens.
Prior to Fedora 4, IDs in Hydra resembled Fedora identifiers ('namespace:id'), and in Sufia we had two opinions about this: 1) the namespace should never be shown to users; 2) the default ID in Fedora 3, a sequential integer, should be overridden by a configurable NOID. In Fedora 4, IDs no longer have a namespace, and so rather than continue to have two kinds of identifiers floating around the system, we have just one: the object identifier, which takes the form of a NOID. This PR eliminates nearly all references to noids in the code (since they're just object identifiers), and makes usage of NOIDs configurable. If you set Sufia.config.enable_noids to false, you will see Fedora 4's UUIDs instead of NOIDs.