Fix permissions when setting visibility to public from authenticated.
Going from authenticated to public used to leave the authenticated
permission objects behind and add the public one as well. This
wasn't a problem with authentication since public is more permissive
than authenticated anyway. But some UI code would get confused and
think the document visibility is both public and authenticated.
Sufia in particular would check the authenticated check box in the
permissions edit form instead of public, which was clearly wrong.