Fix sorting issues with collections.
Uses system_modified_dtsi and system_create_dtsi for sorting instead
of date_uploaded_dtsi and date_modified_dtsi. This applies to both
collections and generic files.

Fix sorting and paging form in my/collections to point to back to
my/collections instead of my/files.

Also remember the collection id when selecting add files to collection
from collection page so that the right collection can be automatically
selected in the collection list.