Provide CanCan abilities for discover permissions
hydra-access-controls provides CanCan abilities for `:read`, `:edit`,
and `:download` permissions, but not `:discover`.  This change add
abilities for `:discover`.

The implementation parallels that used for `:read` and `:edit`
permissions.  It assumes that if a document can be `:read`, then it can
also be `:discover`ed, but not vice-versa.

I added/updated unit tests accordingly.  `ability_spec.rb` has a
comment that references some request specs, but I couldn’t find those
in the codebase so I didn’t add any of those specs.