WorkShowPresenter#file_presenters only returns FileSets
This prevents members that are not FileSets from drawing in the section
of the page that is the list of attached files. For example if you have
a Book that has members that are Sections and FileSets we want to
exclude the Sections from displaying in the table of files.

Ref https://github.com/curationexperts/heliotrope/issues/84