PermissionsController should use Sufia layout
Fixes #1731

This takes a workaround approach, introducing a small amount of technical debt (ticketed in #1815), namely overriding CC's PermissionsController via copypasta. Note that this approach is not new -- this commit is essentially reverting an earlier commit that introduced this bug.

Also: remove unused redlock cruft and add a .rb extension to a related view spec that hasn't been getting run.