Fields displayed on search results come from config
Instead of requiring implementers to override a view template to change
the values that display on the search results page, use the blacklight
configuration as it was intended. This will pave the way for dynamic
configuration features in the future.

There are also a few changes here that make the views more agnostic
about the model type by moving generic_work partials default partials

Fixes #1947