Extracting method to reduce method length
The `CitationsBehaviors::Formatters::OpenUrlFormatter#format` method
was overly long due to the field_map being defined inline. By
extracting the method, we improve the legibility and separate the
different levels of abstraction.

There is also a change in which I switched the `next` guard condition.
There was an assumption that we had a single or an array object, but
then wrap in an array. By moving the logic into an `Array.wrap` method
we tidy up the guard conditions.

Also, note that `Array(Time.now)` explodes the Time object into an
array of its constituent parts:

```ruby
Array(Time.now)
=> [1, 41, 16, 4, 5, 2016, 3, 125, true, "EDT"]
```

So prefer `Array.wrap` when we are using ActiveSupport.