Dump unused Redis config value.
The `config.redis_namespace` value is largely vestigial in CC at this point; it is used only by a migrate task that turns ChecksumAuditLog entries about GenericFiles into ones about FileSets. CC has no legacy data with GenericFiles, unless we are supporting a Curate- or Worthwhile-to-CC path, for which I am seeing a lack of tooling and documentation. This provides a cleaner boundary between Sufia, which cares about Redis for activity streams where namespacing is relevant, and CC, which uses Redis only as a locking mechanism.