Refactoring method usage
There is a bit of disconnect referencing @file_set within the code
and #curation_concern as the interface. Attempting to both consolidate
those concepts and allow for a method to be overridden for file_set/
curation_concern and still allow the internal behavior to work.

In sharing code via a module, I believe it is a good idea to isolate
the usage of an instance variable to a method. This allows for objects
that extend the module to change the accessor method and keep the
existing behavior.