Refactor ingest workflow to use current working file
Fixes projecthydra/sufia#2553.

If the current working file used by the initial ingest is already available locally, it should pass that file to successive jobs. Each subsequent job
can verify the file still exists locally and continue to use it; otherwise, it will download a new local working copy.

Modifications are also made to IngestFileJob so that a Hydra::Derivatives::IoDecorator is always passed to Hydra::Works::AddFileToFileSet with a value for `original_name` and a mime type, if it is provided.