Add workflow tab to administrative sets edit
Fixes projecthydra-labs/hyku#455

The workflow selection was previously part of a dropdown selection on the "Description" tab within the edit Administrative Sets. This request moves workflow selection into a separate tab, allowing the workflow to be linked to an administrative set via radio buttons.

Changes proposed in this pull request:

* add new "workflow" tab with radio buttons
* remove dropdown workflow selection from "description" test
* add specs for new form
* add specs to test the workflow method, to ensure that future changes do not break the form

https://github.com/projecthydra-labs/hyku/issues/455