Use Jbuilder to encode users as JSON
When using the `respond_to` block in the controller the default behavior
for rendering JSON is to use ActiveSupport to serialize the objects. By
removing the `respond_to` block we change the behavior to look for
templates to render the view. Enter Jbuilder.

By using Jbuilder rather than an ActiveSupport serializer we can render
different encodings of User objects in different contexts. It is also
easy to override in implementations of Hyrax like Hyku.