Don't use action_name in i18n routes
The common pattern for a form that fails validation is to have the
create or update method re-render that form. By doing that you have to
have duplicate i18n keys (e.g. edit & update) for a single form. This
change ensures we only need a single i18n value for each.

Fixes #3058