Adding PermissionTemplate.has_many :workflows
This is a concession that each AdminSet has many workflows, but given
that AdminSet is an ActiveFedora::Base and Sipity::Workflows are an
ActiveRecord::Base, we can't cross the streams. So instead, I'm
leveraging the fact that each AdminSet has one and only one Permission
Template and each Permission Template belongs to one and only one
AdminSet.