Call ingest_file synchronously (not in a job)
This enables the ImportURLJob to fetch the remote file and have it
ingested in one job (rather than 2).  If we do it in two jobs there is
the risk that the tempfile is not available to the second job if it runs
on a different machine.

Fixes https://github.com/projecthydra-labs/hyku/issues/745