Fix admin set visibility radio buttons
* Resolves projecthydra-labs/hyku#744
* Resolves #463
* Specs now test for valid release option combinations, and return
appropriate notifications if the selections aren't correct.
* Radio buttons are now checked based on currently saved options.
Related issue #476 allowed the Visibility options to be correctly
checked, but the Release Varies option needed an explicit statement
to correctly know when to show the radio button as checked.

<%= f.radio_button :release_period, '', checked: !f.object.release_varies.blank? %>

This is required because the button is based on simply a value of blank
(because release period can contain either "before" or "embargo"),
so the determining factor as to whether to check it relies on the
existence of one of the underlying release_varies sub-options being
selected. This can be removed if #489 is completed to persist all of
the permission template data.

I was unable to create a failing test for this radio button issue.
Future clean-up of the admin set & permission templates should
hopefully allow a simpler structure that can be validated.