Add field to store whether a workflow allows setting access grants
Sufia allows depositors to set other users who can edit a work.  This is
perfect for a zero step workflow, but for a mediated deposit workflow,
we don't want anyone to be able to edit the work once it is accepted.

We need a place to store whether we should be showing the UI widget that
allows the creator to share access.

Ref https://github.com/projecthydra/sufia/issues/3134