Replacing ordered parameters with named parameters
Given that the first parameter (e.g. `ids`) is an Array, I can see it
being rather easy to instead pass `build_presenters('1', '2', '3')`
which in turn would mean `ids == '1'`, `klass == '2'`, and `args == '3'`

By shifting to named parameters, I hope this clears up the likelihood
of that confusion and unexpected behavior.