Rename ChecksumAuditLog `version` attr to `checked_uri`
This matches what it actually is in the code currently.

The FixityCheckJob says "This URI could include the actual resource (e.g. content) and the version to fixity check... but it could also just be [a file URL]"

It isn't neccesarily a version. It IS always created as a URI -- even when
it's a version, it's the version uri, not the version label. What it actually
is, is the URI that was passed to ActiveFedora::FixityCheckService to check,
which is fine.

Accurate names make code less confusing, trying to make it clearer
what this code does.