Change data from integer to boolean in db-independent way
previous migration from #984, postgres did not like.

This way could be slow, as it needs to fetch and save each object individually,
but I can't think of anything better.

Those who already generated this hyrax app into a local app may have to manually
copy the migration. If this hadn't been in a release yet, that wouldn't seem
like a problem -- but seems it made it into 1.0 somehow?

Fixes #1035

As a bonus, it's reversible now.