Add new feature toggle allowing disabling account creation
When account creation is disabled, redirect to the homepage with an alert. When enabled, let Devise do its thing. This implementation relies on a small routing change that lets Hyrax override two methods in the Devise RegistrationsController. I am on the fence as to whether Hyrax should automatically handle this for users (i.e., we should move the code from the test_app_generator to the install_generator), or if we should rely on documentation and release notes for this. There is a chance that this flipper will be made available in instances that have not made this routing change, which would be misleading for users.

Refs samvera-labs/hyku#1203