Disable terrible rubocop rule: Style/AsciiComments
Firstly, it is culturally exclusionary for no valid reason.

Secondly, some of the most important code for data integrity will be
around UNICODE handling.  The idea that you shouldn't be able to write a
YARD `@example` or even a regular comment using the relevant characters in
question makes this rule fundamentally useless.