Batch create should have same sharing behavior as single
The batch create hadn't yet been fixed so that a share tab only displays
when an admin set that has direct deposit is chosen. We don't want the
depositor to share edit access on something that goes through mediated
deposit.

Prior to this change the batch upload screen could get two sharing tabs.

Ref https://github.com/samvera-labs/hyku/issues/1195