Rework IngestFileJob into FileActor
A major change necessitated by `IngestFileJob`'s reliance on local path
information.  Jobs do not get to pick what worker services them or when.
So the logic cannot depend on the caller prepping a local file in temp
for the job to execute upon. The caller can supply the path as a hint,
like "Use this cached content IF it exists for you", but it cannot be
guaranteed (must be retrievable by other means/identifiers).

Notable changes:
- Use `Hydra::Derivatives::IoDecorator`
- sync/async boolean goes away, since async was not fully supported
- More robust testing
- `FileActor` ONLY acts upon the relation it was initialized with, no
  more backdoor kludgey overrides.  If you want a different relation,
  just get a new actor.  They're cheap.

Not 100% done yet, but this is the bulk of required changes.