Cut over to JobIoWrapper
`JobIoWrapper` is persistence-backed, allowing it to be passed to
asynchronous jobs.

In actors, where async job is spawned, return it (instead of `true`), in case
caller wants to wait for completion or otherwise track/reference it.
This also signifies the difference between "action successfully completed"
like from `.save` and "a deferred thing will determine success/failure".

Replace one-off `after_update_content` callback with `after_perform`
block.  Otherwise, the callback can happen **before** the async event.

In general, I propose that the rails framework callbacks (especially on
ActiveJob) are a universally better approach than our bespoke magic in
`config/initializers/hyrax_callbacks.rb`, giving us:
- `before` callbacks, symmetric to each `after` callback
- `around` callbacks, combining `before` and `after` actions
- a complete range of other callbacks
- the ability to use asynchronicity without breaking timing logic

If the problem is that some job callers want the callback to be
optional while others do not, see the implementation of `:notification`
in `IngestJob`'s `after_perform` block for an example that handles that.

TODO:
- update downstream jobs to accept `JobIoWrapper` for completeness
- update last two jobs using `Hyrax::WorkingDirectory` and remove it