Handle resource_type # => nil case in SolrDocumentBehavior#itemtype
`SolrDocumentBehavior#itemtype` tries to access `resource_type`. By default,
this is fine, but `resource_type` is allowed to return `nil` in client
apps. This handles this case by using an empty array as the default
resource_type value within the method.

`ResourceTypesService#microdata_type` handles the `nil` case, so `[].first` is
an acceptable parameter. The `#microdata_type` behavior is tested, and this PR
adds documentation making it more clear to clients.