Refactoring to more compact methods
Rubocop has a 14 line method length enforcer. These two files exceeded
that method length. By refactoring the methods they are below the 14
line method length.

In the case of the `batch_uploads_controller.rb` the refactoring
introduces a seam in the code that allows Hyrax implementors to more
easily change the response behavior when batches are disabled.

I believe we should aspire to a 10 line method length as the general
rule, with exceptions for data mapping of attributes and other similar
methods.