Refactoring FixityCheckJob
Prior to this refactor, the FixityCheckJob had knowledge of when to
short-circuit the prune ChecksumAuditLog history. By moving the create
and prune directives to ChecksumAuditLog the specs for FixityCheckJob
become much more focused on the FixityCheckJob behavior instead of
the behavioral interactions with the ChecksumAuditLog.

Looking at methods that are at the limit of the Rubocop method length.