Reducing redundant specs
Prior to this commit, the query logic was tested both in the presenter
and in the underlying query service object.

This commit now verifies that the presenter calls the query service
but does not re-test (and re-build the test conditions) of the details
of the query service implementation.

The refactor drops all of LDP requests on the presenter and speeds up
the isolated tests by about 6 seconds. See below for more details:

Before:

```console
Examples with the most LDP requests
  Hyrax::AdminStatsPresenter#works_count when start date is provided queries by start date
    Total LDP: 35 {"HEAD"=>3, "GET"=>22, "POST"=>7, "DELETE"=>2, "PUT"=>1}
  Hyrax::AdminStatsPresenter#works_count when start and end date is provided queries by start and date
    Total LDP: 35 {"HEAD"=>3, "GET"=>22, "POST"=>7, "DELETE"=>2, "PUT"=>1}
  Hyrax::AdminStatsPresenter#works_count includes files but not collections
    Total LDP: 35 {"HEAD"=>3, "GET"=>22, "POST"=>7, "DELETE"=>2, "PUT"=>1}
  Hyrax::AdminStatsPresenter#top_formats gathers formats
    Total LDP: 46 {"HEAD"=>7, "GET"=>27, "POST"=>9, "DELETE"=>2, "PUT"=>1}
  Hyrax::AdminStatsPresenter#works_count when there is uncommitted work provides accurate count, ensuring that solr deletes have been expunged first
    Total LDP: 49 {"HEAD"=>5, "GET"=>31, "POST"=>10, "DELETE"=>2, "PUT"=>1}
  Hyrax::AdminStatsPresenter#active_users returns statistics
    Total LDP: 63 {"HEAD"=>10, "GET"=>37, "POST"=>13, "DELETE"=>2, "PUT"=>1}
Finished in 6.88 seconds (files took 6.87 seconds to load)
```

After:

```console
Examples with the most LDP requests
  [NONE]
Finished in 0.30383 seconds (files took 6.38 seconds to load)
```