Nudging JS wait times and testing sequence
Instead of removing the sleep time, rely instead of Capybara's wait
time. Hyrax is encountering regular failures on finding this content as
part of the Travis build. Hopefully this tweak can assist in reducing
the number of Capybara induced failures in the Travis specs.

Incrementing from 5 to 15 to see if we can prevent erratic failures in
Travis. These failures result in a broken build and require a restart
of the specs. This restart and waiting for feedback slows the overall
trajectory of iterating on Hyrax.