Work form should display selected collections based on params
This passes on values given to the "my" controller (from the collection "Add Works" button), which are currently lost when selecting a work type. Adds a new method to the work form that selects one or more collections (based on the IDs passed in params) automatically in the Relationships tab of the rendered form. This is working for both single works and batches of works.

 * Add method to presenter that adds the `add_works_to_collection` param to the new work/batch path only if it exists.
 * Rename existing param to reflect it is acting on works, not files
 * Fix label of member_of_collection_ids input so i18n matches

Fixes samvera-labs/hyku#1374