Adding auto-wrapping of all Capybara matchers
Based on Notre Dame's experience with our [QA Tests][1], we found it
helpful to wrap all of the Capybara find and matches with sleep
intervals. This reduced the fragility of our feature testing by
augmenting Capybara's semi-reliable `wait` algorithm with sleep
intervals in cases where the `wait` was not enough.

The nature of `Capybara::Rumplestiltskin` is such that most of the time
the specs will be as fast as they can be (they're slow enough as you
are leveraging a browser for this testing). However, when a failure
occurs, `Capybara::Rumplestiltskin` assumes that its due to a slowness
in the browser (and not an actual failure). From there, it sleeps for
a bit, then tries again.

[1]:https://github.com/ndlib/QA_tests/blob/3e1372377f3c019790284196d380860f32c7321a/spec/spec_support/inject_sleep.rb