Share button should link to My Works
When an unauthenticated user clicks the Share Work button on the homepage, they should not be taken to the root dashboard page after logging in, because that is not where works are shared. Instead, redirect the user to the My Works page where the controls for creating works live.

Fixes #1092