Add participants to a new work created directly in a collection
In the save new work actor stake, collection membership is processed the same.  It also determines if there is one and only one collection selected for the work.  If there is, it adds the single collection back into the Environment as :collection_id.  Later, when permissions are being applied from the admin set, it checks for :collection_id in the Environment and adds collection participants for the identified collection.