Citations controller should decide layout like other controllers do
Fixes curationexpects/nurax#40

Otherwise the citation link -- which is publicly viewable -- uses the dashboard layout, which is not used for public users. Using the `decide_layout` strategy like in other Hyrax controllers, though the branch that renders the `dashboard` layout is not currently used. Work to render dashboard-baed work show views is ticketed, so I leave this code as is for consistency and as a foundation for upcoming work.