Stubbing out `signed_in?`
Prior to this commit, and with changes for this branch, the specs were
failing because we introduced a check for `signed_in?`; which in turn
required a Warden::Proxy inclusion. For these tests we don't care about
whether someone is signed in or not (that concern is handled in other
specs).

Below is the exception that was raised.
```console
ActionView::Template::Error:
Devise could not find the `Warden::Proxy` instance on your request environment.
Make sure that your application is loading Devise and Warden as expected and that the `Warden::Manager` middleware is present in your middleware stack.
If you are seeing this on one of your tests, ensure that your tests are either executing the Rails middleware stack or that your tests are using the `Devise::Test::ControllerHelpers` module to inject the `request.env['warden']` object for you.
```