Allow configuration of the predicate used for AdminSet inclusion
Allow a custom predicate to be used to relate objects to their administrative
sets. The default remains `dcterms:isPartOf`. This is the first step in a
process to change the default predicate for this relation. A follow-up PR, to be
merged after the Hyrax 2.0.0 release, will provide warnings for changed usage
and add migration tooling.

This addresses "Phase 1" of #1461 as decscibed at
https://github.com/samvera/hyrax/issues/1461#issuecomment-331205110.