Clean the repository before running tests
These tests depend on the AdminSet existing and having associated
PermissionTemplates.  While the AdminSet did exist, the database cleaner
has removed any associated PermissionTemplates, so the easiest way to
get them recreated was to remove the default admin set, by clearing the
repository.

This also reverts a bad change regarding which attributes should be
passed in.