Updating `samvera-nesting_indexer` gem to v1.0.0
This change updates to `samvera-nesting_indexer` v1.0.0 and addresses
the breaking upstream change for that release. See
[v1.0.0 release notes][release_notes] for details.

The driver for this change is to begin indexing the
`deepest_nested_depth` of an object. This is a necessary attribute as
it helps provide guards to prevent users from nesting objects that
would result in a graph deeper than the maximum allowed depth.

This commit also includes two Rubocop appeasements:

1. Switching from "Style/FileName" to "Naming/FileName"
2. Temporarily disabling the "Lint/UnusedMethodArgument" check

For the first appeasement (e.g. "Naming/Filename"), this addresses and
cleans up an STDOUT warning written by Rubocop.

For the second appeasement (e.g. "Lint/UnusedMethodArgument"), this
addresses the implementation details and constraints of the
`it_behaves_like 'a Samvera::NestingIndexer::Adapter'` RSpec
declaration, of which Hyrax::Adapters::NestingIndexAdapter adhears to.

Related to #2221

[release_notes]:https://github.com/samvera-labs/samvera-nesting_indexer/releases/tag/v1.0.0