Controller that adds members to a collection
This allows members to be added to a collection from Dashboard -> Works -> batch Add Collection -> select a collection with deposit access.  You cannot use the regular controller for Dashboard -> Collections -> edit collection because it enforces edit access on the collection, as it should.  To be able to add a member, the user only needs deposit access.