Allow `representative_media` partial calls without a `viewer` flag
The partial now treats calls without a `viewer` flag as `viewer:
false`. Applications may be calling this partial without the flag, which has
worked previously, so this improves compatibility.

We might also consider adding a warn level log in this case. It will often (but
perhaps not always) be the case that this indicates an out of date show view
attempting to render the partial.