Prevent circular nesting
Relates to #2221

The nested collection query service needs to remove collections from
the lists for nesting if they would cause nested indexing errors.
This updates the query service to exclude collections which would cause
circular indexing errors. It also prevents nesting of collections which
would exceed the total nesting depth.

Additionally this implements a modification to the collection factory
which allows for creating a solr_document for a nestable collection for
querying purposes, without requiring the full collection creation and
nested indexing process.

The rules for allowed nesting are as follows:

* if adding as a parent collection:
1) cannot have any pathnames containing the collection's ID
2) cannot already be Collection G's direct parent
3) cannot cause total nesting depth to exceed limit

* if adding as a child (subcollection):
1) Cannot have any pathnames containing any of the child's pathname or ancestors
2) cannot already be the collection's direct child
3) cannot cause total nesting depth to exceed limit