Avoid potential infinite loop in `WorkShowPresenter`
The `WorkShowPresenter#representative_presenter` can enter an infinite loop in
the case that it returns another instance of itself as representative. This
behavior has been spotted this in production shortly after upload of a new item,
presumably while background jobs are running, when visiting the work show page
immediately after deposit. This clears itself up shortly after, and doesn't
present a major issue. Still, it seemed worth addressing the edge case.

Note that an infinite loop is still technically possible within the method; we
believe this fixes the real case we have encountered.