Support new collections features for old collections_id actor syntax
We readded support for `:member_of_collection_ids` syntax in
`CollectionsMembershipActor`, but several behaviors have been added in the
meanwhile. This adds support for those.

Namely, `#extract_collection_id` is updated to handle the old syntax, which
ensures that `#create` populates the `:collection_id` environment attribute in
the same way as it does for the new sytax. Likewise, `#valid_membership?` is
updated to accept the collection ids it should check; existing calls are changed
to pass the ids directly.

For a default collections configuration, these methods shouldn't impact the
behavior of existing Hyrax 2.0.x actor stack calls using the
`:member_of_collection_ids` syntax. We may wish to introduce integration tests
for this, since the crucial behavior here is spread across several actors and
depends on their interaction.