Use analytics fixture in analytics spec tests
The current `analytics_spec` class assumes that there are no changes
made to the config file generated by engine_cart. The Analytics team is,
however, needing to run local google analytics and as a result of
updating this file tests are failing.

Changes proposed in this pull request:
* Add an analytics.yml fixture which mirrors the engine_cart generated
yml
* allow the tests to use the fixture file rather than the current file
on disk