add single/multi-membership spec tests
Right now, all tests allow the work to be added to both collections even when it shouldn’t.  These tests can be used to determine when the problem is fixed.

There is one tests for single membership that is incomplete, set to skip, and marked with a TODO because I wasn’t sure how to get capybara to do the collection selection process on the work’s edit form Relationships tab.