Wire up remove nesting buttons
Add remove buttons to the parent collections and subcollections list
on the collection show page.

With the move of the remove options to the show page, the relationships
tab is now unnecessary as all options are available on the show page.
The relationships tab being available only to those with edit authority
was problematic, as it didn't give appropriate access to the
relationship management options.

Make the buttons functional to remove a parent collection or a
subcollection.