add facet for collections
This is not ideal.  It would be better if Admin Sets where listed in the same filter as the Collections.  But creating that filter is very complex and would likely require a kludge because Admin Sets and Collections are not stored in the same solr field.

Also, it would be even better if the facet was a pivot on collection type and collection/admin set name.  But that is even more complex.

For Hyrax 2.1 release, I recommend going with this implementation.  For a later release, we can explore putting in a more complex filter.