Replace save with reindex on child collection
References https://github.com/samvera/hyrax/issues/2798

The nested child persistence service updates the member_of_collections
relationship between a child and parent. Active Fedora makes all of the
expected changes to the data in Fedora with no need for a subsequent
save. However, it doesn't cause a subsequent reindex on the child
collection to record the relationship in Solr.

Since no save is required to the child, this PR is changing the save to
simply force reindexing. To be as efficient as possible, it calls the
nested reindexer directly, rather than using update_index which would
trigger the additional call to the reindexer, thereby reindexing the
child collection twice.

Note:
In the case of a new member_of_collections relationship, Active Fedora
does, however, save and reindex the parent (unnecessarily because
nothing is changing on the parent!). It doesn't do this for removal of
the relationship.