Add partial nested indexing option
Fixes https://github.com/samvera/hyrax/issues/2798

While this doesn't solve the issue of the unnecessary save of the parent
collection when adding a subcollection or work to the collection, it
does solve the problematic results of that save (the nested reindexing
of the parent's entire graph).

Saving the parent collection loses its nested indexing fields, so the
reindexing is necessary for the parent collection. But by adding
a parameter to the nesting indexer, this parameter can be referenced by
the adapter to allow the children of this parent to be reindexed only if
necessary (only when they are missing THEIR nested indexing fields).

By removing the reindexing of all child collections and child works,
this change will remove most of the issues with in-line indexing of
nested collections.

Limited reindexing will be used for:
- indexing of collection when adding a work to a collection
- indexing of parent collection when adding a subcollection
- indexing a collection update where no nesting changes take place

(includes additional changes as needed for samvera-nesting_indexer v2.0.0)